Flask-App/requirements.txt
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-auth==2.40.3
googleapis-common-protos==1.70.0
httplib2==0.30.0
idna==3.10
oauthlib==3.3.1
proto-plus==1.26.1
protobuf==6.32.0
pyasn1_modules==0.4.2
pyasn1==0.6.1
pyparsing==3.2.3
pytz==2025.2
requests-oauthlib==2.0.0
requests==2.32.5
rsa==4.9.1
uritemplate==4.2.0
urllib3==2.5.0
asgiref==3.9.1
blinker==1.9.0
click==8.2.1
colorama==0.4.6
DateTime==5.5
dotenv==0.9.9
flask-cors==6.0.1
Flask==3.1.2
greenlet==3.2.4
gspread==6.2.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
playwright==1.55.0
pyee==13.0.0
python-dotenv==1.1.1
setuptools==80.9.0
typing_extensions==4.15.0
Werkzeug==3.1.3
zope.interface==7.2